# src/Core/System/Payment/OpenApi/Resources/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Contena Payment API",
        "version": "1.0.0",
        "description": "Provider-neutral payment, query, refund, subscription and transfer operations. Monetary amounts use the currency's minor unit."
    },
    "servers": [
        {
            "url": "/payment-api/v1"
        }
    ],
    "paths": {
        "/pay": {
            "post": {
                "operationId": "pay",
                "summary": "Create a payment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/SignedRequest"
                                    },
                                    {
                                        "type": "object",
                                        "required": ["external_order_no", "amount", "method_code", "subject"],
                                        "properties": {
                                            "external_order_no": {"type": "string", "maxLength": 64},
                                            "amount": {"type": "integer", "minimum": 1},
                                            "currency_code": {"type": "string", "minLength": 3, "maxLength": 3, "default": "CNY"},
                                            "method_code": {"type": "string", "maxLength": 32},
                                            "subject": {"type": "string", "maxLength": 255},
                                            "channel_code": {"type": "string", "maxLength": 32},
                                            "device_type": {"type": "string", "maxLength": 32},
                                            "client_ip": {"type": "string", "maxLength": 64},
                                            "notify_url": {"type": "string", "format": "uri", "maxLength": 2048},
                                            "return_url": {"type": "string", "format": "uri", "maxLength": 2048},
                                            "channel_extra": {"type": "object", "additionalProperties": true}
                                        }
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {"$ref": "#/components/responses/PaymentSuccess"},
                    "400": {"$ref": "#/components/responses/PaymentError"},
                    "401": {"$ref": "#/components/responses/PaymentError"},
                    "422": {"$ref": "#/components/responses/PaymentError"}
                }
            }
        },
        "/query": {
            "post": {
                "operationId": "queryPayment",
                "summary": "Query a payment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {"$ref": "#/components/schemas/SignedRequest"},
                                    {
                                        "type": "object",
                                        "properties": {
                                            "order_no": {"type": "string", "maxLength": 64},
                                            "external_order_no": {"type": "string", "maxLength": 64}
                                        },
                                        "description": "Provide order_no or external_order_no."
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {"$ref": "#/components/responses/PaymentSuccess"},
                    "400": {"$ref": "#/components/responses/PaymentError"},
                    "401": {"$ref": "#/components/responses/PaymentError"},
                    "404": {"$ref": "#/components/responses/PaymentError"}
                }
            }
        },
        "/refund": {
            "post": {
                "operationId": "refundPayment",
                "summary": "Create a refund",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {"$ref": "#/components/schemas/SignedRequest"},
                                    {
                                        "type": "object",
                                        "required": ["external_refund_no", "refund_amount"],
                                        "properties": {
                                            "order_no": {"type": "string", "maxLength": 64},
                                            "external_order_no": {"type": "string", "maxLength": 64},
                                            "external_refund_no": {"type": "string", "maxLength": 64},
                                            "refund_amount": {"type": "integer", "minimum": 1},
                                            "reason": {"type": "string", "maxLength": 255}
                                        },
                                        "description": "Provide order_no or external_order_no."
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {"$ref": "#/components/responses/PaymentSuccess"},
                    "400": {"$ref": "#/components/responses/PaymentError"},
                    "401": {"$ref": "#/components/responses/PaymentError"},
                    "404": {"$ref": "#/components/responses/PaymentError"},
                    "409": {"$ref": "#/components/responses/PaymentError"},
                    "422": {"$ref": "#/components/responses/PaymentError"}
                }
            }
        },
        "/subscribe": {
            "post": {
                "operationId": "subscribe",
                "summary": "Create a recurring agreement",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {"$ref": "#/components/schemas/SignedRequest"},
                                    {
                                        "type": "object",
                                        "required": ["external_subscription_no"],
                                        "properties": {
                                            "external_subscription_no": {"type": "string", "maxLength": 64},
                                            "channel_code": {"type": "string", "maxLength": 32},
                                            "notify_url": {"type": "string", "format": "uri", "maxLength": 2048},
                                            "return_url": {"type": "string", "format": "uri", "maxLength": 2048},
                                            "period_type": {"type": "string", "maxLength": 16},
                                            "period": {"type": "integer", "minimum": 1},
                                            "execute_time": {"type": "string", "format": "date-time"},
                                            "single_amount": {"type": "integer", "minimum": 1},
                                            "total_amount": {"type": "integer", "minimum": 1},
                                            "total_payments": {"type": "integer", "minimum": 1},
                                            "channel_extra": {"type": "object", "additionalProperties": true}
                                        }
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {"$ref": "#/components/responses/PaymentSuccess"},
                    "400": {"$ref": "#/components/responses/PaymentError"},
                    "401": {"$ref": "#/components/responses/PaymentError"},
                    "422": {"$ref": "#/components/responses/PaymentError"}
                }
            }
        },
        "/transfer": {
            "post": {
                "operationId": "transfer",
                "summary": "Transfer funds",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {"$ref": "#/components/schemas/SignedRequest"},
                                    {
                                        "type": "object",
                                        "required": ["external_transfer_no", "amount", "payee", "payee_name"],
                                        "properties": {
                                            "external_transfer_no": {"type": "string", "maxLength": 64},
                                            "amount": {"type": "integer", "minimum": 1},
                                            "currency_code": {"type": "string", "minLength": 3, "maxLength": 3, "default": "CNY"},
                                            "payee": {"type": "string", "maxLength": 128},
                                            "payee_name": {"type": "string", "maxLength": 64},
                                            "channel_code": {"type": "string", "maxLength": 32},
                                            "remark": {"type": "string", "maxLength": 255},
                                            "notify_url": {"type": "string", "format": "uri", "maxLength": 2048},
                                            "channel_extra": {"type": "object", "additionalProperties": true}
                                        }
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {"$ref": "#/components/responses/PaymentSuccess"},
                    "400": {"$ref": "#/components/responses/PaymentError"},
                    "401": {"$ref": "#/components/responses/PaymentError"},
                    "409": {"$ref": "#/components/responses/PaymentError"},
                    "422": {"$ref": "#/components/responses/PaymentError"}
                }
            }
        },
        "/notify/{channel}/{channelConfigId}": {
            "post": {
                "operationId": "receiveProviderNotification",
                "summary": "Receive a payment provider notification",
                "parameters": [
                    {"name": "channel", "in": "path", "required": true, "schema": {"type": "string"}},
                    {"name": "channelConfigId", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[0-9a-f]{32}$"}}
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {"schema": {"type": "object", "additionalProperties": true}},
                        "application/x-www-form-urlencoded": {"schema": {"type": "object", "additionalProperties": true}}
                    }
                },
                "responses": {
                    "200": {"description": "Provider-specific acknowledgement"},
                    "400": {"$ref": "#/components/responses/PaymentError"},
                    "404": {"$ref": "#/components/responses/PaymentError"},
                    "422": {"$ref": "#/components/responses/PaymentError"}
                }
            }
        }
    },
    "components": {
        "schemas": {
            "SignedRequest": {
                "type": "object",
                "required": ["app_id", "timestamp", "sign"],
                "properties": {
                    "app_id": {"type": "string", "maxLength": 64},
                    "timestamp": {"oneOf": [{"type": "integer"}, {"type": "string", "pattern": "^[0-9]+$"}]},
                    "nonce": {"type": "string", "maxLength": 128},
                    "sign": {"type": "string", "pattern": "^[0-9a-f]{64}$"}
                },
                "description": "Sign all non-empty top-level parameters except sign. Sort names ascending, join as name=value with ampersands, append &key={app_secret}, then calculate a lowercase HMAC-SHA256 using app_secret. The timestamp must be within 300 seconds of server time."
            },
            "PaymentResult": {
                "type": "object",
                "required": ["status", "action"],
                "properties": {
                    "resource_no": {"type": "string", "description": "Contena resource number"},
                    "external_resource_no": {"type": "string", "description": "Calling app's resource number"},
                    "transaction_no": {"type": "string", "description": "Contena transaction number when applicable"},
                    "status": {"type": "string", "enum": ["unknown", "pending", "processing", "succeeded", "failed", "closed"]},
                    "action": {"type": "string", "enum": ["none", "redirect", "html", "qr_code", "client"]},
                    "action_value": {"type": "string", "nullable": true}
                }
            },
            "SuccessEnvelope": {
                "type": "object",
                "required": ["code", "message", "data"],
                "properties": {
                    "code": {"type": "string", "enum": ["SUCCESS"]},
                    "message": {"type": "string", "enum": ["ok"]},
                    "data": {"$ref": "#/components/schemas/PaymentResult"}
                }
            },
            "ErrorEnvelope": {
                "type": "object",
                "required": ["code", "message", "data"],
                "properties": {
                    "code": {"type": "string"},
                    "message": {"type": "string"},
                    "data": {"nullable": true}
                }
            }
        },
        "responses": {
            "PaymentSuccess": {
                "description": "The operation was accepted and mapped to a unified payment result.",
                "content": {
                    "application/json": {"schema": {"$ref": "#/components/schemas/SuccessEnvelope"}}
                }
            },
            "PaymentError": {
                "description": "The request was rejected.",
                "content": {
                    "application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}
                }
            }
        }
    }
}